+++ azure-cli2017.pyproj
@@ -247,7 +247,7 @@
     <Compile Include="command_modules\azure-cli-ams\azure\__init__.py" />
     <Compile Include="command_modules\azure-cli-ams\azure_bdist_wheel.py" />
     <Compile Include="command_modules\azure-cli-ams\setup.py" />
-    <Compile Include="command_modules\azure-cli-appservice\azure\cli\command_modules\appservice\azure_devops_build_iteractive.py" />
+    <Compile Include="command_modules\azure-cli-appservice\azure\cli\command_modules\appservice\azure_devops_build_interactive.py" />
     <Compile Include="command_modules\azure-cli-appservice\azure\cli\command_modules\appservice\azure_devops_build_provider.py" />
     <Compile Include="command_modules\azure-cli-appservice\azure\cli\command_modules\appservice\commands.py" />
     <Compile Include="command_modules\azure-cli-appservice\azure\cli\command_modules\appservice\custom.py" />
